NewsFeedService/requirements.txt
# News Feed System - Python Dependencies

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database & ORM
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1

# Caching & Session
redis==5.0.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# HTTP Client
httpx==0.25.2
requests==2.31.0

# Async Task Queue
celery==5.3.4
kombu==5.3.4

# CORS Support
python-cors==1.0.1

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1